move into dag resolution instead of control flow
